Make another attempt at fixing packaging.
The prior attempt at fixing packaging appeared to work, but was due to
some stale generated files that ended up in an include path. This fixes
the packaging properly.

STATIC_ROOT wasn't actually meshing well with the include path. The
include path now points to the "static" directory instead of
"static/djblets", and STATIC_ROOT now expands to an empty string. This
allows an import like "@{STATIC_ROOT}/djblets/" to match "djblets/"
within Djblets's "static/" directory.

This also means that, while we can safely use STATIC_ROOT for imports,
we cnanot use it for anything that will expand within the .css output
(such as image URLs).