Fix filtering by reviewer in the issue summary table.
The redesign of the issue summary table had a last-minute change that
converted an attribute on an element into a data attribute, and a rename
of a function, which broke filtering of reviewers. I've fixed this, and
changed an element lookup to use a stored reference to an element for
the filter widget.

Testing Done:
Tested filtering with two users. "All Users" and each user entry worked
properly.

Reviewed at https://reviews.reviewboard.org/r/7530/