Update some of the screenshots in the user manual.
We've refreshed much of the UI in Review Board 2.5 up until this point,
some of it quite recently. As such, many of the screenshots in the user
manual are very much out-of-date.

This provides a new set of screenshots the comment dialog, file
attachment thumbnails, issue summary table, and review request boxes.
It also improves some of the content in these boxes, highlighting rich
text in the comment dialog and additional thumbnail types for file
attachment thumbnails.

Not everything has been updated, but enough of the major changes have.
The rest will go in after beta 2.

Testing Done:
Built the docs and looked over the screenshots in desktop and mobile
mode.

Reviewed at https://reviews.reviewboard.org/r/7532/