Fix filename caption rendering
Previously, captions were escaped twice: once with `escapejs` in the
`reviewable_page_data.js` template and again via the underscore template
in the `RB.FileAttachmentThumbnail`. Since the only place a
`RB.FileAttachmentModel` will retrieve an escaped value for the caption
is from the JS template, we unescape it there so that the model recieves
the correct value.

The filename is now also correctly parsed.

Testing Done:
Ran JS tests.

Verified that a caption with unsafe characters was correctly set in
both the JS model and the database model.

Reviewed at https://reviews.reviewboard.org/r/7538/