Opening and closing rich text editors no longer creates drafts
Previously, opening and closing a Markdown editor would mark the field
as dirty and, in the case of the review request editor, would create a
new review request draft. The reason this occurs is that CodeMirror's
`isClean` does not work the way we want (or it lies), so we check for
dirtiness ourselves by comparing the initial value with CodeMirror's
reported value.

Testing Done:
Ran JS tests.

Verified that editing either the text or the "Enable Markdown" setting
creates a draft while just opening and closing the editor does not.

Reviewed at https://reviews.reviewboard.org/r/7539/