The e-mail backend is always reset in tests
Previously, if a test using the `console_email_backend` decorator
raised an exception, the e-mail backend would not be reset to the old
backend, causing other tests to fail. This is no longer the case.

Testing Done:
Ran unit tests.

Tests that used this decorator that raised an exception caused other
tests to fail without this patch applied but did not with it applied.

Reviewed at https://reviews.reviewboard.org/r/7562/