Show send e-mail controls in diffviewer
Previously, the send e-mail controls were not showing in the diffviewer
becuase the template context lacked the "send_email" context variable.
This has been corrected and the controls now render properly in the
diffviewer when e-mail is enabled.

Testing Done:
Ran unit tests.

Verified the controls show when e-mail is enabled and do not when it is
disabled.

Reviewed at https://reviews.reviewboard.org/r/7566/