Change syntax.css to be a documented .less file.
syntax.css was our only CSS file (outside of 3rd-party CSS files) that
was still a plain CSS file, which made it hard to clean up generated CSS
files from the static directory. It also added a barrier to future
improvements.

This is now a .less file, and no longer appears to be minified. It's
documented, and ready to be extended to style CodeMirror syntax classes.

Testing Done:
Checked some diffs and rich text. Looked the same as it did before.

Reviewed at https://reviews.reviewboard.org/r/7540/