Update CodeMirror to 5.5, and add new syntax modes.
We've been on CodeMirror 3.23, and attempted an upgrade to 4.2 in the
past. That upgrade resulted in infinite loops and display problems, so
we abandoned it. It seems though that CodeMirror 5.5 is pretty stable
for our needs, and has the added benefit of mobile support, so we're
switching over as a trial run.

This was also build with additional syntax highlighter modes for
CoffeeScript, CSS, Go, HTML, Perl, PHP, Python, ReStructuredText, Ruby,
shell scripts, SQL, XML, and YAML. This will help with providing a
better preview of code when writing using Markdown triple-backticks.

Along with these additions, I've turned off the `elasticChars` switch,
which attempted to match up braces to the correct indentation level. This
wasn't working in practice for us, for some reason, and it's best not to be
too much like an IDE when people are expecting a typical text field.

Testing Done:
Tested this in Chrome, Firefox, and iOS. I didn't see any of the
lockups or anything we had before (which may be due to fixes in
CodeMirror, or fixes in our event handling, which may have previously
triggered infinite loops).

I tested each of the new syntax modes, and they seem to work well.

Reviewed at https://reviews.reviewboard.org/r/7541/