Attempt to standardize syntax styles between Pygments and CodeMirorr.
Syntax highlighting appeared very differently between Pygments and
CodeMirror, with the latter using its default theme. While a nice
default, it sharply changed when saving the text.

We now have a greater degree of consistency between the two. It is
definitely not perfect, as the two have different ideas as to the
granularity they should use for determining style types, and some simply
conflict.

Still, this is a lot better than what we had, and the differences are
minor enough. CodeMirror now mostly follows what we're using for
Pygments.

Testing Done:
Tested a bunch of code from different languages and compared the results.
Things looked nearly identical, with some exceptions (such as `True` in
Python, which we can't make the same without messing up other things).

Reviewed at https://reviews.reviewboard.org/r/7542/