Fix display of Gravatars in several places.
We were using the raw `siteconfig.settings.get()` call to figure out
whether we want to use Gravatars. This wasn't able to fall back to the
default, since that's just a simple dictionary function. By using
`siteconfig.get()` instead, we're able to fall back to the default,
showing Gravatars on installs where the checkbox hasn't been toggled off
and back on.

This also fixes a parameter order issue when building the review reply
list. The Gravatar argument was in the wrong location, resulting it it
always being on. This had a worse side-effect of the comment ID not
being set. This meant if a user had a draft reply and reloaded the page,
they would not be able to update, publish, or delete the comment.

Testing Done:
Verified that Gravatars were working in all locations.

Verified that I could save/delete draft replies again after a page reload.

Reviewed at https://reviews.reviewboard.org/r/7543/