Fix the spacing above tabs in the text review UIs with tabs.
The text review UI pages containing tabs (like the Markdown page) had an
ugly gap between the review request action bar and the rest of the box.
This was due to how a margin for the tabs was interacting with other
elements on the page. Switching that margin to being a padding fixes
this.

When not using tabs, the header was clashing a bit with the tab color,
making for a bad transition. In this case, we now keep the standard
header color.

Testing Done:
Tested a variety of file attachments. The text-based ones looked
conssitent within their type of file now, and there was no ugly
gap.

Reviewed at https://reviews.reviewboard.org/r/7547/