ReviewReplyDraftBannerView now properly removed
Previously the `ReviewReplyDraftBannerView` was not correctly removed
when a review reply was published; the element would be removed from
the DOM, but the event listeners would still fire (and cause errors in
the console). Now, the view is correctly removed with
`Backbone.View.remove` so the elements are removed from the DOM and all
event listeners are disconnects.

Testing Done:
Ran JS tests.

With this patch applied, the error no longer appears in the console.

Reviewed at https://reviews.reviewboard.org/r/7564/