Add documentation for webhooks and custom payloads.
This change adds a new page to the admin manual for the webhooks UI, focusing
mostly on custom payloads. This is additionally linked to from the custom
payload form in the admin UI so that users won't be mystified when it comes to
choosing what to type into the box.

Testing done:
- Built docs HTML and looked at the output. Checked links.
- Checked the admin page to make sure the link went to the right place.

Reviewed at https://reviews.reviewboard.org/r/7591/