Fix the user page and filter-related UI issues on mobile.
The user page's tabs weren't flush with the datagrid, due to the filter
wrapping below it on mobile. This created a pretty nasty look, and it
wasn't the only place the filters were messing up the display on mobile
(depending on the mobile device).

A couple things were done to address this:

1. The tabs for the user page no longer include the username in the
   tab title (since that can be inferred from the page), saving
   precious room, especially with long usernames.

2. The filters are now kept in a menu, activated when clicking a filter
   icon. This icon docks to the right of the datagrid title area. When
   clicking, a menu pops up with a little fade-in animation, offering
   the possible filters on the page. (We may want to move to this
   later for desktop as well.)

While fixing this, I ended up creating a base template for datagrids
that handled tabs and filters in a consistent manner.

Testing Done:
Tested in iPhone 4/6 screen sizes, and on desktop. The filters all worked,
and the user page was viewable at all screen sizes.

Reviewed at https://reviews.reviewboard.org/r/7545/