Prevent the diff viewer from styling .txt files.
Pygments assumes that .txt files should be styled using ResourceLexer by
default, when guessing by filename. This is very often wrong, and
results in a bunch of error indicators in the style, for practically
every word in the file.

We now maintain a blacklist of filename extensions that won't be styled.
It so far just contains .txt files.

Testing Done:
Loaded a text file in the diff viewer before the fix, and saw all the
error indicators.

Loaded a text file after the fix and saw that it looked like a plain
text file.

Reviewed at https://reviews.reviewboard.org/r/7623/