Fix 500 errors when two review requests depend on each other.
If two review requests are marked as depending on each other, serializing
things to the API would fail due to the circular reference. The issue here is
that the draft model on the frontend was requesting that the depends on field
be expanded.

The only reason it was needing to be expanded was because it wanted to get the
"url" field. Since we already had custom serializers for the links in the
review request model, I've extended these a bit to add the URL field (and
copied them to the draft resource), so we don't need to fetch the fully
expanded dependency object.

Testing done:
Set two review requests to depend on each other. Verified that I could now make
changes to the drafts without getting 500 errors.

Fixes bug 3955.

Reviewed at https://reviews.reviewboard.org/r/7643/