Fix extra borders on age columns in datagrids.
The datagrid CSS was supposed to remove the right-side border for the last
column, but it wasn't because we had an important rule in the age columns that
was overriding it. We also had an extra border on the top of the first row.
This change fixes both of those issues.

Testing done:
Looked at a datagrid with a few review requests. Saw that the extra border was
gone.

Reviewed at https://reviews.reviewboard.org/r/7648/