Ensure SSH is given a valid port number.
When an invalid port is given in a repository path, a cryptic error message was shown.

The change provides a better error message to the user when an invalid port is in the repository path.

Testing Done:
Ran unit tests.

Manually verified that the error message was displayed upon replication.

Bugs closed: 3891

Reviewed at https://reviews.reviewboard.org/r/7619/