Fix the DiffCommitIndexView to no longer use hoverIntent
The hoverIntent jQuery plugin was removed as a dependency on the
`release-2.5.x` branch and that change was eventually merged into the
`dvcs` branch. However, the `DiffCommitIndexView` utilized this plugin.
The view has been reworked to no longer depend on it and a bit of
cleanup has occurred.

Testing Done:
Verified that the summaries expanded in the diff viewer.

Reviewed at https://reviews.reviewboard.org/r/7652/