On ReviewRequest "Depends On" field, show discarded dependencies with strikethrough
The "Depends On" field on review requests showed submitted
dependencies with strikethrough to indicate it is no longer
relavant, but not for discarded dependencies.

Modified the html rendering for the field to also show
the discarded dependencies with strikethrough.

Testing Done:
 - Tested on local dev server.
 - Passes /reviews test

Bugs closed: 3758

Reviewed at https://reviews.reviewboard.org/r/7635/