Changed the group reviews header to use the group name
Originally, on the dashboard of the group it would say "Review requests for group.name"

It now uses the groups display name in this string instead of group.name. The motivation for this change is that group display names are often more verbose and meaningful.

Testing Done:
Tested on Chrome and Firefox to verify that the header displays correctly.

Reviewed at https://reviews.reviewboard.org/r/7649/