Improve file attachment hover logic
Previously, if a user put their mouse in just the right place (to the
right of a thumbnail), all thumbnails would fade. This was because the
condition for a thumbnail to not fade was for the mouse to be hovering
over the `.file` inside. This issue has been solved by selectively
fading out thumbnails and not fading out the entire `#file-list`
container.

This patch also replaces the use of a `self = this` binding in
JavaScript with `_.bind(..., this)`.

Testing Done:
Ran JS tests.

Manually verified that placing the mouse cursor near the thumbnails no
longer triggers the old behaviour of fading out all attachments.
Instead, the currently hovered over attachment does not fade out but
all other do.

Reviewed at https://reviews.reviewboard.org/r/7673/