Only send public flag when pubishing reviews from banner
Publishing from the `RB.DrafReviewBannerView` now only sends the
`public` flag to the server during the save call. This fixes a bug where
publishing through the banner would sometimes result in text being
rendered as raw HTML.

This occurred because the `RB.ReviewDialogView` would save the
`body_top` and/or `body_bottom` text and request HTML back so that it
could be rendered. The JS model would recieve this and save it. Then,
when publishing, the banner would send the entire model along, including
the `body_top` and/or `body_bottom` as HTML, but saying that it is
Markdown. Thus the server would end up with a bunch of HTML.

Testing Done:
The bug no longer reproduces with this patch applied. Instead, reviews
are published as intended (without HTML markup).

Reviewed at https://reviews.reviewboard.org/r/7678/