Add a trivial_publish capability
Since not all versions of Review Board will support trivial publishes,
it is a good idea to expose this is a server capability. That way,
`rbtools` and other consumers of the API can determine if publishing
(e.g., with `rbt publish`) can be done trivially.

Reviewed at https://reviews.reviewboard.org/r/7681/