Ensure that e-mails always include absolute URLs for image comments.
The comment thumbnails (selected regions) for image file attachments are
included in e-mail notifications for the review, but when using the built-in
file storage backend, the URLs given to these are relative to the server. This
change makes it so if the returned URL is relative, we prefix it with the
server URL.

Testing Done:
I used the email backend so that I could print emails to the console. Then I
posted a review with an image attached and posted a review for the image. Also,
I ran the unit tests and it passed them.

Bugs closed: 3944

Reviewed at https://reviews.reviewboard.org/r/7688/