Ensure default reviewers are added when creating post-commit review requests.
When creating a review request for a committed change via the "New Review
Request" page, default reviewers were not added to it. This was happening
because the code to add those reviewers was only called when posting changes
from RBTools.

To fix this, in the review request manager, we now explicitly add default
reviewers when creating a review request from a committed change.

Testing done:
- Created a new review request from the web UI and saw that default reviewers
  were correctly added.
- Ran unit tests.

Bugs closed: 3900

Reviewed at https://reviews.reviewboard.org/r/7691/