Add support for Subversion and Perforce Assembla repositories.
This introduces the ability to add Subversion and Perforce repositories
hosted on Assembla. This service also offers Git, but lacks an API for
properly communicating with Git repositories.

Assembla requires a couple of specific Perforce settings, namely a
P4HOST matching the given Assembla project name, and a P4CHARSET of
"utf8".

The Perforce support had to be updated to take a host, and the
repository form had to be updated to give the hosting service more
control of how data could be added to the repository.

The Subversion support was updated to use the credentials from the
hosting service, if available.

Testing Done:
Unit tests pass.

Manually tested Perforce and Subversion repositories.

Reviewed at https://reviews.reviewboard.org/r/7703/