Do not hide footer text field when it isn't empty
Normally, when the review dialog is open, if there are no comments the
footer text field will be hidden. However, it is possible for a user to
create some comments, add a footer, and then delete the comments. When
that user opens the review dialog again, the footer will be hidden and
thus uneditable.

We no longer hide the footer text field when it has text. Additionally,
if while the review dialog is open the footer text is cleared and there
are no open comments, it will be hidden.

A unit test has been added to test this new behaviour and ensure it does
not regress. Other unit tests have been updated to account for this new
behaviour.

Testing done:
- Ran JS tests.
- Created a review with top and bottom text. Clearing the bottom text
  caused the footer text editor to disappear.

Testing Done:
- Ran JS tests.
- Created a review with top and bottom text. Clearing the bottom text
  caused the footer text editor to disappear.

Reviewed at https://reviews.reviewboard.org/r/7743/