Change linenum color of diff viewer for better user visibility.
The color in diff viewer before was black by default, and on many browsers,
rendered with a bold weight. This was distracting from the code, which is much
more important than the line numbers. This change makes the line number color
lighter and a normal weight.

Testing done:
Reloaded the page to see if the color had changed.

Fixes bug 3981.

Reviewed at https://reviews.reviewboard.org/r/7687/