Fix Summary column compatibility with Python 2.6.
The Summary column in 2.5 switched to using `format_html()`, but was using
it in a way that was incompatible with Python 2.6. This wasn't caught
due to the lack of unit tests.

We now use specific positional arguments (`{0}, {1}`) instead of generic
arguments (`{}`), in order to retain compatibility.

Unit tests were added for all variations on the column's rendered
output.

Testing Done:
Unit tests pass on Python 2.6 and 2.7.

Bugs closed: 4005

Reviewed at https://reviews.reviewboard.org/r/7747/