Fix presentational issues with image diff mode selection.
When initially viewing an image diff, the mode will be set to Two-Up,
but that mode won't appear selected in the UI, due to the code setting
the class name being nested in a conditional that only executed if this
wasn't the first mode set. That has been fixed.

When selected, though, the navigation would end up wrapping, due to how
the default width was computed. We now compute the width correctly,
factoring in padding/margins, to prevent wrapping.

Testing Done:
Tested in Chrome and Firefox. Saw that Two-Up was selected by default, and
that it didn't wrap.

Reviewed at https://reviews.reviewboard.org/r/7748/