Make standard auth backend tests more reliable
This patch no longer assumes the `StandardAuthBackend` is in a
particular place in the list of enabled authentication backends.
Instead, we now iterate through the enabled backends and return the
first one that is an instance of `StandardAuthBackend` (and not an
instance of a subclass of it).

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7756/