Defining a mode for imageAttachmentView
When rendering the image review view, the div class for the image itself
in imageAttachmentView mode currently has name `image-diff-null`. This
is because the base class `BaseImageView` initializes `mode` to `null`
but the subclass `imageAttachmentView` does not define a proper name for
it. This can potentially confuse devs, or at the very least, it gives
the impression that the class was not extended properly.

This patch fixes the problem by setting the `mode` to `attachment` so
that the div class will be `image-diff-attachment` rather than
`image-diff-null`.

Testing Done:
Used Inspect Element tool on the browser to ensure that the div class
now reads `image-diff-attachment`.

Reviewed at https://reviews.reviewboard.org/r/7772/