Advertise GitHub Enterprise support in Power Pack
This patch adds support for fake hosting services that Review Board
does not provide, but Power Pack does. These services do not implement
anything, but are instead shown in the "Add Repository" form when Power
Pack is not enabled.

When a fake hosting service is selected in the form, an advertisement
for Power Pack is shown and the form becomes unsubmittable. This is to
ensure that users don't end up getting an error when trying to use the
hosting service that is not actually registered in the database.

A large portion of this change is updating `repositoryform.js` to bring
it in line with the rest of our JavaScript by prefixing all jQuery
objects with `$`.

Testing Done:
Ran unit tests.

Manually verified that the fake hosting service appeared only when
Power Pack was not enabled.

Reviewed at https://reviews.reviewboard.org/r/7757/