Fix a couple of bugs when using condensediffs.
The condensediffs command had a couple of regressions that weren't
caught before release.

The first is that our loop to find RawFileDiffData objects was, upon
encountering an IntegrityError, attempting to look up the entry using
the wrong hash. We wanted to use the hash provided on that instance, but
were using a different value, leading to a DoesNotExist error.

The second is that our percentage for the progress bar was overly
accurate, showing the entire floating point number. It now shows an
integer value.

Testing Done:
Ran condensediffs. Saw that the percentage was correct, and that we weren't
crashing anymore upon hitting IntegrityErrors.

Reviewed at https://reviews.reviewboard.org/r/7773/