Allow changing review request ownership
The owner of a review request can now be changed by either the review
request's current owner (i.e., the submitter) or a superuser. This
allows review requests to be transferred from one user to another if
they've been abandoned so that the history of the change is not lost.

The API for review request drafts now accepts a `submitter` field and
the UI has been updated so that the `submitter` field is editable. Like
the `target_people` field, setting a submitter that does not exist will
result in an error and the field will be reset.

Unit tests have been added for this new behaviour.

Testing done:
- Ran unit tests
- Ran JS tests.

Reviewed at https://reviews.reviewboard.org/r/6990/