Add user field to change descriptions
The `ChangeDescription` model now supports recording a user that is
associated with the change. The user should be retrieved via the
`ChangeDescription.get_user` method, which will delegate determining the
user for the change description to the model the change is associated
with if it has not yet been determined. Once determined, it will be
saved to the database for faster retrieval in the future.

The `user` field can currently only be calculated for change
descriptions associated with review requests.

Testing Done:
- Ran unit tests.
- Viewed a review request that had its owner changed before this patch
  was applied. The user was determined correctly for the change
  descriptions.

Reviewed at https://reviews.reviewboard.org/r/7765/