Advertise TFS support in Power Pack
Add support for fake SCMTools entries that Review Board does not
provide, but Power Pack does. These tools are not implemented; they
simply appear in the Tool dropdown in the "Add Repository" form when
Power Pack is not enabled.

When a fake tool is selected in the form, an advertisement for Power
Pack is shown and the form becomes unsubmittable. This is to ensure
that users don't end up getting an error when trying to use a tool that
is not actually registered in the database.

This patch also moves all of the tool information serialization into
the `RepositoryForm` instead of the JS template.

Testing Done:
Ran unit tests.

Manually verified the following:

- The form cannot be submitted when a fake tool is selected.
- The Power Pack advertisement shows when a fake tool is selected.
- The fake SCM Tool entries do not appear when Power Pack is installed.

Reviewed at https://reviews.reviewboard.org/r/7758/