Change the format of our README to ReStructuredText.
Now that we're posting official releases to PyPI, our README is being
shown to any users visiting the website. This does not support Markdown,
only ReST. In order to show something sane, we need to change the format
of our README. Fortunately, GitHub is happy enough with most ReST files.

The content of this remains the same, with the addition of Assembly
repositories. The transition to ReST is pretty sane. The only thing
worth pointing out is that we're using named references for links when
it's a proper name for the site we're linking to or otherwise makes
sense ("RBTools documentation", etc.), and inline references where it
felt more natural.

Testing Done:
Looked at the README as rendered HTML. Checked links.

Reviewed at https://reviews.reviewboard.org/r/7759/